requirements.txt
instaloader==4.9.5
python-dotenv==0.21.0